@@ -15,7 +15,10 @@
         "psr-4": { "Volcanus\\QueryBuilder\\": "src/" }
     },
     "require": {
-        "php": ">=5.5.9",
-        "volcanus/database": "dev-master"
+        "php": "^5.6||^7.0",
+        "volcanus/database": "^1.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^5.7.20||^6.0"
     }
 }
