@@ -2,7 +2,7 @@
 /**
  * Volcanus libraries for PHP
  *
- * @copyright 2011-2013 k-holy <k.holy74@gmail.com>
+ * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
  */
 
@@ -15,31 +15,31 @@ use Volcanus\QueryBuilder\Adapter\Mysql\MysqlExpressionBuilder;
  *
  * @author k.holy74@gmail.com
  */
-class MysqlExpressionBuilderTest extends \PHPUnit_Framework_TestCase
+class MysqlExpressionBuilderTest extends \PHPUnit\Framework\TestCase
 {
 
-	public function testResultColumn()
-	{
-		$builder = new MysqlExpressionBuilder();
-		$this->assertEquals('name AS `alias`', $builder->resultColumn('name', 'alias'));
-	}
+    public function testResultColumn()
+    {
+        $builder = new MysqlExpressionBuilder();
+        $this->assertEquals('name AS `alias`', $builder->resultColumn('name', 'alias'));
+    }
 
-	public function testAsDate()
-	{
-		$builder = new MysqlExpressionBuilder();
-		$this->assertEquals("DATE_FORMAT(name, '%Y-%m-%d')", $builder->asDate('name'));
-	}
+    public function testAsDate()
+    {
+        $builder = new MysqlExpressionBuilder();
+        $this->assertEquals("DATE_FORMAT(name, '%Y-%m-%d')", $builder->asDate('name'));
+    }
 
-	public function testAsTimestamp()
-	{
-		$builder = new MysqlExpressionBuilder();
-		$this->assertEquals("DATE_FORMAT(name, '%Y-%m-%d %H:%i:%s')", $builder->asTimestamp('name'));
-	}
+    public function testAsTimestamp()
+    {
+        $builder = new MysqlExpressionBuilder();
+        $this->assertEquals("DATE_FORMAT(name, '%Y-%m-%d %H:%i:%s')", $builder->asTimestamp('name'));
+    }
 
-	public function testAsGeometry()
-	{
-		$builder = new MysqlExpressionBuilder();
-		$this->assertEquals("ASTEXT(name)", $builder->asGeometry('name'));
-	}
+    public function testAsGeometry()
+    {
+        $builder = new MysqlExpressionBuilder();
+        $this->assertEquals("ASTEXT(name)", $builder->asGeometry('name'));
+    }
 
 }
